✨: 사용자 정보를 표시하는 UserDialog 컴포넌트 추가 및 사용자 선택 상태 관리 훅 구현
