+++ src/pages/posts-manager/ui/PostsManagerPage.tsx
@@ -34,8 +34,10 @@ import {
 import { HighlightedText } from '@/shared/ui/HighlightedText';
 
 import { useSelectedPostStore } from '@/features/posts/model/useSelectedPostStore';
+import { useSelectedUserStore } from '@/features/users/model';
 import PostAddDialog from '@/widgets/post/ui/PostAddDialog';
 import PostUpdateDialog from '@/widgets/post/ui/PostUpdateDialog';
+import UserDialog from '@/widgets/user/ui/UserDialog';
 import type { Comment, Tag } from '../model/types';
 
 export const PostsManagerPage = () => {
@@ -74,7 +76,7 @@ export const PostsManagerPage = () => {
   const [showEditCommentDialog, setShowEditCommentDialog] = useState(false);
   const [showPostDetailDialog, setShowPostDetailDialog] = useState(false);
   const [showUserModal, setShowUserModal] = useState(false);
-  const [selectedUser, setSelectedUser] = useState<User | null>(null);
+  const setSelectedUser = useSelectedUserStore((state) => state.setSelectedUser);
 
   // 게시물 관련 훅
   const {
@@ -609,35 +611,7 @@ export const PostsManagerPage = () => {
       </CustomDialog>
 
       {/* 사용자 모달 */}
-      <CustomDialog open={showUserModal} onOpenChange={setShowUserModal} title='사용자 정보'>
-        <img
-          src={selectedUser?.image}
-          alt={selectedUser?.username}
-          className='w-24 h-24 rounded-full mx-auto'
-        />
-        <h3 className='text-xl font-semibold text-center'>{selectedUser?.username}</h3>
-        <div className='space-y-2'>
-          <p>
-            <strong>이름:</strong> {selectedUser?.firstName} {selectedUser?.lastName}
-          </p>
-          <p>
-            <strong>나이:</strong> {selectedUser?.age}
-          </p>
-          <p>
-            <strong>이메일:</strong> {selectedUser?.email}
-          </p>
-          <p>
-            <strong>전화번호:</strong> {selectedUser?.phone}
-          </p>
-          <p>
-            <strong>주소:</strong> {selectedUser?.address?.address}, {selectedUser?.address?.city},{' '}
-            {selectedUser?.address?.state}
-          </p>
-          <p>
-            <strong>직장:</strong> {selectedUser?.company?.name} - {selectedUser?.company?.title}
-          </p>
-        </div>
-      </CustomDialog>
+      <UserDialog open={showUserModal} onOpenChange={setShowUserModal} />
     </Card>
   );
 };
